error handling for API

+++ api/queries/collections.py
@@ -21,7 +21,12 @@ class CollectionIn(BaseModel):
     fields: List[Field]
 
 
+class CollectionUpdate(BaseModel):
+    fields: List[Field]
+
+
 class CollectionOut(CollectionIn):
+    owner: str
     id: str
 
 
@@ -30,13 +35,14 @@ class CollectionListOut(BaseModel):
 
 
 class CollectionRepository:
-    def create_collection(self, collection: CollectionIn) -> CollectionOut:
+    def create_collection(self, owner, collection: CollectionIn) -> CollectionOut:
         new_collection = {
             "name": collection.name,
+            "owner": owner,
             "fields": [field.dict() for field in collection.fields],
         }
-        if db.collections.find_one({"name": collection.name}):
-            raise HTTPException(status_code=404, detail="collection name in use")
+        if db.collections.find_one({"name": collection.name, "owner": owner}):
+            raise HTTPException(status_code=403, detail="collection name in use")
         db.collections.insert_one(new_collection)
         new_collection["id"] = str(new_collection["_id"])
         return CollectionOut(**new_collection)
@@ -49,28 +55,41 @@ class CollectionRepository:
         except:
             raise HTTPException(status_code=404, detail="invalid collection_id")
 
-    def delete_collection(self, collection_id: str):
+    def delete_collection(self, current_user: str, collection_id: str):
         try:
-            result = db.collections.delete_one({"_id": ObjectId(collection_id)})
-            if result.deleted_count > 0:
-                return {"message": "collection deleted"}
+            collection = db.collections.find_one({"_id": ObjectId(collection_id)})
         except:
             raise HTTPException(status_code=404, detail="invalid collection_id")
+        if collection["owner"] != current_user:
+            raise HTTPException(status_code=401, detail="not authorized to delete specified collection")
+            result = db.collections.delete_one({"_id": ObjectId(collection_id)})
+        if result.deleted_count > 0:
+            return {"message": "collection deleted"}
+        
 
     def update_collection(
-        self, collection_id: str, collection: CollectionIn
+        self, current_user: str, collection_id: str, collection_update: CollectionUpdate
     ) -> CollectionOut:
         try:
-            result = db.collections.update_one(
-                {"_id": ObjectId(collection_id)}, {"$set": collection.dict()}
-            )
-            return self.get_collection(collection_id)
+            collection = db.collections.find_one({"_id": ObjectId(collection_id)})
         except:
             raise HTTPException(status_code=404, detail="invalid collection_id")
+        if collection["owner"] != current_user:
+            raise HTTPException(status_code=401, detail="not authorized to update specified collection")
+        collection_update = {
+            "name": collection["name"],
+            "owner": collection["owner"],
+            "fields": [field.dict() for field in collection_update.fields]
+        }
+        result = db.collections.update_one(
+            {"_id": ObjectId(collection_id)}, {"$set": collection_update}
+        )
+        return self.get_collection(collection_id)
+        
 
-    def get_list_of_collections(self) -> CollectionListOut:
+    def get_list_of_collections_by_owner(self, username) -> CollectionListOut:
         collections = []
-        db_cursor = db.collections.find()
+        db_cursor = db.collections.find({"owner": username})
         for collection in db_cursor:
             collection["id"] = str(collection["_id"])
             collections.append(collection)
